fix(web): unify browser supabase client for auth diagnostics
